boost/test: reset test tree enabled/disabled state at start of unit_test_main()
boost::test has a static tree of tests that has nodes marked as
enabled or disabled based on the command-line arguments.  In
order to permit launching unit_test_main() multiple times, it
is necessary to reset this state at the start of each invocation.

This commit extends the test_tree_visitor to conditionally permit
visiting disabled nodes and uses this pattern to reset the test
state at launch.